Update regources and go

See merge request tokend/horizon!33

+++ web_v2/requests/get_fee_list.go
@@ -44,7 +44,7 @@ type GetFeeList struct {
 		Subtype     int64            `fig:"subtype"`
 		FeeType     int32            `fig:"fee_type"`
 		Account     string           `fig:"account"`
-		AccountRole int32            `fig:"account_role"`
+		AccountRole uint64           `fig:"account_role"`
 		LowerBound  regources.Amount `fig:"lower_bound"`
 		UpperBound  regources.Amount `fig:"upper_bound"`
 	}
